Abandon Cordova-based attempt
It turns out that services cannot be implemented in TypeScript and the workarounds (e.g. notifications) work only if the application is started and running in the background.

Services could be implemented in Java and then integrated as a plugin. This approach however defeats much of the main purpose of using this project to learn TypeScript.