Change the minimal time amount to 2
This is mainly driven by the observation that "in 1 day" was displayed from the moment a duration went below 2 days until it went below 1 day. IOW, the time left could almost halve while the same duration is shown, which is a bit strange.

As an added benefit, we now no longer need resources for singular and plural units.

References #24